Refactor price storage to cents and fix PDF service - Issue #62

+++ simplyFact/database/migrations/2026_04_13_091028_create_meals_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
                 ->constrained()
                 ->cascadeOnDelete();
             $table->integer('number_of_meal');
-            $table->float('total_price');
-            $table->float('reimbursed_price');
+            $table->unsignedInteger('total_price');
+            $table->unsignedInteger('reimbursed_price');
             $table->timestamps();
         });
     }
